Prevented memoryleaks due to implicit conversion
Removed implicit conversion from GameObject* to GmObjctPtr
This should prevent two GmObjctPtr's pointing to the same GameObject but not knowing about eachother

+++ KrokEngine/KrokEngine/Testing/MovingObject.hpp
@@ -6,7 +6,7 @@ class MovingObject :
     public ImageGameObject
 {
 private:
-    double i;
+    float i;
 
 public:
     MovingObject(float x, float y);
